Split extra_cpp tests into sound and unsafe
